+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## dev
+- add query `round_model_performances` (#60)
+
 ## [2.9.0] - 2021-09-15
 - support passing `round_num` to `list_datasets`, to get available files from past rounds
 - `download_dataset` no longer requires a destination path, it defaults to the source file name